% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsSpec.R
\name{cohByStruct}
\alias{cohByStruct}
\title{Subset Frequencies Through Postwhitening}
\usage{
cohByStruct(coh, maxFreqOffsetIdx, sigfx, sigfy)
}
\arguments{
\item{coh}{A \code{matrix} containing coherence between one predictor
and the response.}

\item{maxFreqOffsetIdx}{An \code{integer} representing exactly what it seems.}

\item{sigfx}{A \code{vector} containing postwhiten-based significant frequencies
of the input.}

\item{sigfy}{A \code{vector} containing postwhiten-based significant frequencies
of the response.}
}
\description{
Zeroes out non-significant variance contribution
}
