% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/coherence.R
\name{coherence}
\alias{coherence}
\title{Calculates the coherence between two series.}
\usage{
coherence(x, y, method = c("aveOfMsc", "mscOfAve", "mscOfAveCoh"), nw = 4,
  k = 7, nFFTx = NULL, centre = "none", dtx = 1, dty = 1,
  blockSizex = NULL, overlap = 0, freqRange = NULL, maxFreqOffset = 0,
  convertMscToNormal = FALSE, namex = "x", namey = "y")
}
\arguments{
\item{x}{slower sampled series - the response (stupid)}

\item{y}{faster sampled series - the predictor}

\item{method}{only "aveOfMsc" is currently implemented}

\item{nw}{blap}

\item{k}{blap}

\item{centre}{The time series is centred using one of three methods:
expansion onto discrete prolate spheroidal sequences ('Slepian'),
arithmetic mean ('arithMean'), trimmed mean ('trimMean'), or not at all ('none').}
}
\description{
Multitaper coherence estimate.
}
\details{
Only calculates the forward coherence currently.  'x' is the response if
the series do not have the same sampling rate ... which is confusing.
}
