% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{df.std}
\alias{df.std}
\title{Standardize a data.frame}
\usage{
df.std(x)
}
\arguments{
\item{x}{The \code{data.frame} whose columns will be standardized.}
}
\description{
Subtract mean and divide by sd for each column of the provided data.frame.
}
