% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsSpec.R
\name{eigenCoef}
\alias{eigenCoef}
\title{Calculate weighted eigencoefficients}
\usage{
eigenCoef(x, nw = 4, k = 7, nFFT = "default", centre = "none",
  deltat = 1, dtUnits = "second", adaptiveWeighting = TRUE,
  dpssIN = NULL)
}
\description{
Calculates weighted eigencoefficients for a single vector / block
}
