% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsSpec.R
\name{eigenCoefFit}
\alias{eigenCoefFit}
\title{Returns fitted values for eigencoefficients}
\usage{
eigenCoefFit(H, Hinfo, ykx, prednames, centFreqIdx)
}
\arguments{
\item{H}{A \code{matrix} with the transfer function.}

\item{Hinfo}{All info reltaed to H, a list probably ...}

\item{ykx}{An \code{array} containing the predictor eigencoefficents}

\item{prednames}{I don't think this even gets used...}

\item{centFreqIdx}{mmmmm...}
}
\description{
DO NOT USE THIS FUNCTION, use predictEigencoef()
}
