% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsSpec.R
\name{filter.twoSided}
\alias{filter.twoSided}
\title{Apply a linear filter to a time-series}
\usage{
filter.twoSided(x, filter)
}
\arguments{
\item{x}{A \code{vector} containing the series to filter.}

\item{filter}{A \code{vector} containing the filter coefficients.  See details.}
}
\description{
Assumes a two-sided filter.
}
\details{
Ends up being twice as fast as stats::filter().  Only 2-sided.  Only convolution.

The argument, filter, should be of length 2*L + 1 (i.e., odd).  The positive lag (causal) filter
coefficients should be in elements (L+1):(2*L) with the negative lag (acausal) filter coefficients
contained in the 1:L elements of filter.  The L+1 index of filter is lag-0.  This is the same convention as
required by \code{stats::filter()}.
}
