% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsSpec.R
\name{fixDeadBand}
\alias{fixDeadBand}
\title{this only works for a very specific situation}
\usage{
fixDeadBand(msc, zeroOffsetIdx, freqRangeIdx, band, df, replaceWith = 0)
}
\description{
this only works for a very specific situation
}
