% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{msc2norm}
\alias{msc2norm}
\title{Convert the MSC to Standard Normal}
\usage{
msc2norm(msc, k, J = 1, mu = 0, sd = 1, avoidInf = FALSE)
}
\arguments{
\item{msc}{A \code{matrix} containing MSC values.}

\item{k}{An \code{integer} representing the number of tapers used in MSC calculation.}

\item{J}{An \code{integer} indicating the number of blocks used in MSC calculation.}

\item{mu}{The mean of the Standard Normal.}

\item{sd}{The standard deviation of the Standard Normal.}

\item{avoidInf}{A \code{logical} indicating whether to avoid potential infinite values.
Returns 10 or -10 instead.

Do NOT change this parameter.  Results are not reliable if changed.}
}
\description{
Convert the MSC to Standard Normal
}
