% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utilsSpec.R
\name{postWhMulti}
\alias{postWhMulti}
\title{A Wrapper for transfer3::postwhiten()}
\usage{
postWhMulti(ykx, dkx, yky, dky, chiSigLev = 0.99, nFreqBand)
}
\arguments{
\item{ykx}{A \code{compelx array} of dimension MxKxP (freqs by tapers by
predictors).}

\item{dkx}{A \code{real array} containing the adaptive weights of ykx.}

\item{yky}{A \code{complex matrix} containing the response eigencoefficients (MxK).}

\item{dky}{A \code{real matrix} containing the adaptive weights of yky.}

\item{chiSigLev}{A \code{numeric} between 0 and 1 representing the
confidence level above which to consider something significant.}

\item{nFreqBand}{An \code{integer} indicating how many frequency bands
should be used to postwhiten.  Each band is postwhitened individually.}
}
\description{
Post-whitens multiple eigencoefficient series
}
\details{
The number of frequency bands can be around 50-ish (based on
conversation with DJT).
}
