% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{predictEigenCoef}
\alias{predictEigenCoef}
\title{Eigencoefficient Prediction}
\usage{
predictEigenCoef(H, newdata = NULL, yk = NULL)
}
\arguments{
\item{H}{A \code{list} returned by tf() containing H, Hinfo, and info.}

\item{newdata}{A \code{data.frame} containing the new data.  Must have the same column names as the original
data used in the transfer function estimation.}

\item{newyk}{A \code{list} of matrices.  One matrix per predictor with the kth column giving the kth
eigencoefficient for that predictor.}
}
\description{
Predicts the eigencoefficients based on an estimated transfer function and new data provided.
}
