% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{predictTs}
\alias{predictTs}
\title{Predict the Time Series from the predictors and impulse response}
\usage{
predictTs(newdata, ir, Hinfo, info)
}
\arguments{
\item{newdata}{A \code{data.frame} containing the predictors.  Column names must match what is
present in Hinfo.}

\item{ir}{A \code{matrix} containing the impulse responses corresponding to the rows in Hinfo.}

\item{Hinfo}{A piece of what is returned by tf().}

\item{info}{Another piece of what is returned by tf().}
}
\description{
Predict the Time Series from the predictors and impulse response
}
