% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{specPredict}
\alias{specPredict}
\title{Spectrum prediction
Predicts the spectrum based on an estimated transfer function and new data provided.}
\usage{
specPredict(H, d2)
}
\arguments{
\item{H}{A \code{list} returned by tf() containing H, Hinfo, and info.}

\item{d2}{A \code{data.frame} containing the new data.  Must have the same column names as the original
data used in the transfer function estimation.}
}
\description{
Spectrum prediction
Predicts the spectrum based on an estimated transfer function and new data provided.
}
