% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{standardizeDf}
\alias{standardizeDf}
\title{Standardizes each column of a data.frame
Standardizes by subtracting the mean and dividing by the standard deviation.}
\usage{
standardizeDf(x)
}
\description{
Standardizes each column of a data.frame
Standardizes by subtracting the mean and dividing by the standard deviation.
}
