% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transfer.R
\name{tf}
\alias{tf}
\title{Calculates some transfer functions.}
\usage{
tf(y, x, nw = 4, k = 7, nFFTy = NULL, centre = "none", dty = 1,
  dtx = 1, blockSizey = NULL, overlap = 0, cohSigLev = 0.99,
  nOffAllowed = 1, forceZeroOffset = TRUE, freqRange = NULL,
  maxFreqOffset = 0, deadBand = NULL, decorrelate = NULL)
}
\arguments{
\item{y}{A single column \code{data.frame} containing the response time-series.}

\item{x}{A \code{data.frame} containing the predictor time-series.}

\item{nOffAllowed}{An \code{integer} representing the number of offsets, \emph{including
the zero-offset}, to be used in the transfer function estimation.}

\item{decorrelate}{A vector containing the names of the predictors (columns of x)
in the order of the iterative regression.}
}
\description{
Calculates some transfer functions.
}
